requirements.txt
black==18.6b4
celery==4.2.0
coverage==4.5.1
dj-pagination==2.3.2
django-activity-stream==0.6.5
django-annoying==0.10.4
django-cachalot==2.0.1
django-contrib-comments==1.8.0
django-countries==5.3.1
django-crumbs==0.5.0
django-environ==0.4.5
django-gravatar2==1.4.2
django-markwhat==1.6.0
django-picklefield==1.0.0
django-recaptcha==1.4.0
django-redis==4.9.0
django-threadedcomments==1.2
Django==1.11.14
djangorestframework==3.8.2
furl==1.2
isort==4.3.4
Markdown==2.6.11
pinax-badges==2.0.0
psycopg2==2.7.5
raven==6.9.0
requests==2.19.1
teerace-tml==0.3
webstack_django_sorting==1.0.1
